ref: Move consts, utils from `integrations/opentelemetry/` (#4272)
Moving stuff out of `integrations/opentelemetry/` step by step since
there is no OpenTelemetry integration anymore -- it's part of the core
SDK. This moves
`sentry_sdk/integrations/opentelemetry/{utils,consts}.py` ->
`sentry_sdk/opentelemetry/{utils,consts}.py`.

There's quite some stuff moving around in this PR in order to remove
circular dependencies that have to do with `tracing.py`:
- `get_span_status_from_http_code` from `tracing.py` is now in
`tracing_utils.py`
- various constants from `tracing.py` are now in `consts.py`

Additionally, Sphinx was unhappy, so tweaked the way we type some things
in `api.py` and `_init_implementation.py`.

I'll possibly follow this up with creating a nice structure for the
tracing files, maybe a common `tracing` directory with `tracing.py`,
`tracing_utils.py` -> `utils.py`, and dedicated `consts.py`.

Ref https://github.com/getsentry/sentry-python/issues/3853
