+++ sentry_sdk/opentelemetry/consts.py
@@ -23,8 +23,6 @@ SPAN_ORIGIN = "auto.otel"
 
 
 class SentrySpanAttribute:
-    # XXX not all of these need separate attributes, we might just use
-    # existing otel attrs for some
     DESCRIPTION = "sentry.description"
     OP = "sentry.op"
     ORIGIN = "sentry.origin"
